+++ include/boost/beast/_experimental/unit_test/suite.hpp
@@ -11,6 +11,7 @@
 #define BOOST_BEAST_UNIT_TEST_SUITE_HPP
 
 #include <boost/beast/_experimental/unit_test/runner.hpp>
+#include <boost/throw_exception.hpp>
 #include <ostream>
 #include <sstream>
 #include <string>
@@ -565,7 +566,7 @@ fail(std::string const& reason)
     if(abort_)
     {
         aborted_ = true;
-        throw abort_exception();
+        BOOST_THROW_EXCEPTION(abort_exception());
     }
 }
 
@@ -583,7 +584,7 @@ suite::
 propagate_abort()
 {
     if(abort_ && aborted_)
-        throw abort_exception();
+        BOOST_THROW_EXCEPTION(abort_exception());
 }
 
 template<class>